refactor: extract 26/WAKU-PAYLOAD related function to waku_payload.ts
